pyproject.toml
[project]
name = "tarotreading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "marimo>=0.13.15",
    "pillow>=11.2.1",
]
